requirements.txt
torch==2.6.*
torchvision==0.21.*

--index-url https://pypi.org/simple
torchmetrics>=1.3.0
tqdm>=4.64
pyyaml>=6.0
numpy>=1.23
pandas>=1.5
matplotlib>=3.7
scikit-learn>=1.2
pillow>=10.0